+++ JesterGameProto/Assets/Scripts/Player/HealthBar.cs
@@ -7,13 +7,13 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxValue(int health)
+    public void SetMaxValue(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
